requirements.txt
ollama>=0.3.0
rich>=13.0.0
python-magic>=0.4.27
PyPDF2>=3.0.0
